feat: introduce shared extension types and refactor related imports
